rename states -> usStates to avoid confusion
